+++ RELEASES.md
@@ -9,6 +9,7 @@
 ### Bug fixes
 
 -   Fixed lag issue while navigating through the menu [[#138](https://github.com/AstraExt/astra-monitor/issues/138)] (related to [GNOME GLib issue #3229](https://gitlab.gnome.org/GNOME/glib/-/issues/3229))
+-   `nvidia-smi` parsing performance improved by ~40%
 -   Fixed CPU frequency not being correctly displayed [[#155](https://github.com/AstraExt/astra-monitor/issues/155)] [[#168](https://github.com/AstraExt/astra-monitor/issues/168)]
 -   Fixed GPU detection on some GPUs [[#162](https://github.com/AstraExt/astra-monitor/issues/162)]
 -   Fixed GPU model name not being correctly displayed [[#152](https://github.com/AstraExt/astra-monitor/issues/152)]